backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-community==0.3.0
chromadb==0.4.24
pypdf==4.3.1
google-generativeai==0.7.2
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
pytesseract==0.3.13
Pillow==10.4.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
